backend/competitive_services_requirements.txt
# Combined requirements for CodeChef + Codeforces services

# Web Framework
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.10.5

# Web Scraping
selenium==4.27.1
beautifulsoup4==4.12.3
webdriver-manager==4.0.2

# HTTP Requests
requests==2.32.3

# AI/LLM SDKs
google-generativeai==0.8.3
openai==1.59.6

# Environment Management
python-dotenv==1.0.1

# HTML Parsing
lxml==5.3.0
